+++ src/main/java/com/mysite/sbb/repository/AnswerRepository.java
@@ -1,4 +1,4 @@
-package com.mysite.sbb;
+package com.mysite.sbb.repository;
 
 import com.mysite.sbb.entity.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
