+++ src/test/java/com/mysite/sbb/SbbApplicationTests.java
@@ -2,6 +2,8 @@ package com.mysite.sbb;
 
 import com.mysite.sbb.entity.Answer;
 import com.mysite.sbb.entity.Question;
+import com.mysite.sbb.repository.AnswerRepository;
+import com.mysite.sbb.repository.QuestionRepository;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
